Create separate component to hold articles
